requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pdfplumber>=0.10.0
easyocr>=1.7.0
Pillow>=10.0.0
google-generativeai>=0.3.0
reportlab>=4.0.0
python-multipart>=0.0.6